Extract Converter class
